refactor(a2ui): switch catalog imports to store/* and react/* paths
Move the catalog `<Name>/index.tsx` files (and `defineCatalog.ts`) to
import from `store/types` and `react/A2UIRenderer` so they sit on the
post-refactor module layout.

Add three thin compatibility re-exports so the new paths resolve
against the still-present `core/` implementation:

- `src/store/types.ts` → re-exports `core/types.js`
- `src/react/A2UIRenderer.tsx` → re-exports `A2UIRender as A2UIRenderer`
  and `NodeRenderer` from `core/A2UIRender.jsx`
- `src/react/useDataBinding.ts` → re-exports `core/useDataBinding.js`

The full headless-renderer refactor (#2536) replaces these shims with
real implementations; this PR only relocates the catalog import surface
so that follow-up rebases shrink to a no-op delete of the shims.
