+++ packages/genui/a2ui/src/catalog/Button/index.tsx
@@ -1,8 +1,8 @@
 // Copyright 2026 The Lynx Authors. All rights reserved.
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
-import { A2UIRender } from '../../core/A2UIRender.jsx';
-import type { GenericComponentProps } from '../../core/types.js';
+import { A2UIRenderer } from '../../react/A2UIRenderer.jsx';
+import type { GenericComponentProps } from '../../store/types.js';
 
 import '../../../styles/catalog/Button.css';
 
@@ -40,7 +40,7 @@ export function Button(
   return (
     <view className='button' bindtap={handleClick}>
       {childResource
-        ? <A2UIRender resource={childResource} />
+        ? <A2UIRenderer resource={childResource} />
         : <text>Button</text>}
     </view>
   );
