+++ packages/genui/a2ui/src/catalog/Column/index.tsx
@@ -1,8 +1,8 @@
 // Copyright 2026 The Lynx Authors. All rights reserved.
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
-import { NodeRenderer } from '../../core/A2UIRender.jsx';
-import type { GenericComponentProps } from '../../core/types.js';
+import { NodeRenderer } from '../../react/A2UIRenderer.jsx';
+import type { GenericComponentProps } from '../../store/types.js';
 
 import '../../../styles/catalog/Column.css';
