+++ packages/genui/a2ui/src/catalog/List/index.tsx
@@ -3,9 +3,9 @@
 // LICENSE file in the root directory of this source tree.
 import type * as v0_9 from '@a2ui/web_core/v0_9';
 
-import { NodeRenderer } from '../../core/A2UIRender.jsx';
-import type { GenericComponentProps } from '../../core/types.js';
-import { useDataBinding } from '../../core/useDataBinding.js';
+import { NodeRenderer } from '../../react/A2UIRenderer.jsx';
+import { useDataBinding } from '../../react/useDataBinding.js';
+import type { GenericComponentProps } from '../../store/types.js';
 
 import '../../../styles/catalog/List.css';
