+++ packages/genui/a2ui/src/catalog/Text/index.tsx
@@ -1,7 +1,7 @@
 // Copyright 2026 The Lynx Authors. All rights reserved.
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
-import type { GenericComponentProps } from '../../core/types.js';
+import type { GenericComponentProps } from '../../store/types.js';
 import '../../../styles/catalog/Text.css';
 
 /**
