+++ packages/genui/a2ui/src/react/useDataBinding.ts
@@ -1,0 +1,5 @@
+// Copyright 2026 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
+// Compatibility re-export. See `react/A2UIRenderer.tsx` for context.
+export * from '../core/useDataBinding.js';
